Support delegate methods without default implementation.
The problem is the following:

There are optional protocol methods which - when implemented - do not
have a valid "default" return value.  Or in other words, if you do
implement the method, then you must do something meaningful in it.

Examples are NSTableViewDelegate's
tableView:dataCellForTableColumn:row: and
tableView:viewForTableColumn:row:.  Both are optional, but if you do
implement them, then you must provide a real implementation and not
return null.  Another example is tableView:heightOfRow: (we return 16
in NSTableView.cs, which may not be correct if the user specified a
very large font in Interface Builder).

This happens in MonoMac when using any .NET event on the view: doing
so sets the view's delegate to an auto-generated implementation of
NSTableViewDelegate.

I also tried calling NSTableView's makeViewWithIdentifier:owner: (in
my native Objective C test app) as it's described in Apple's API docs
for tableView:dataCellForTableColumn:row:, but that doesn't work, that
method returns nil.

The solution:

Well, as trivial as it may sound, the solution is simply not to
implement these methods - and since .NET doesn't allow any
unimplemented methods, we need to trick the Objective C runtime into
thinking they are.  Which is actually as simple as overriding
NSObject's respondsToSelector:

For NSTableViewDelegate, adding the RespondsToSelector() override like
this will do the trick - this is a test:
https://gist.github.com/7e30f0980eee75187d19, we just need to replace
that with a real implementation.

We now generate the following in NSTableView.g.cs:

======
	static IntPtr selGetRowHeight = Selector.GetHandle ("tableView:heightOfRow:");
	static IntPtr selGetDataCell = Selector.GetHandle ("tableView:dataCellForTableColumn:row:");
	static IntPtr selGetViewForItem = Selector.GetHandle ("tableView:viewForTableColumn:row:");
	static IntPtr selRespondsToSelector = Selector.GetHandle ("respondsToSelector:");
	[Export ("respondsToSelector:")]
	bool _RespondsToSelector (Selector sel)
	{
		if (sel.Handle.Equals (selGetRowHeight))
			return getRowHeight != null;
		if (sel.Handle.Equals (selGetDataCell))
			return getDataCell != null;
		if (sel.Handle.Equals (selGetViewForItem))
			return getViewForItem != null;
		return Messaging.bool_objc_msgSendSuper_intptr (SuperHandle, selRespondsToSelector, sel.Handle);
	}
====